requirements.txt
absl-py
appnope
astunparse
attrs
backcall
backports.entry-points-selectable
bayesian-optimization
bleach
blis
boto3
botocore
cachetools
catalogue
certifi
cffi
clang
click
cloudpickle
cycler
cymem
debugpy
decorator
defusedxml
distlib
entrypoints
filelock
first
Flask
flatbuffers
gast
gensim
google-api-core
google-auth
google-auth-oauthlib
google-cloud-bigquery
google-cloud-bigquery-storage
google-cloud-core
google-crc32c
google-pasta
google-resumable-media
googleapis-common-protos
grpcio
gym
h5py
idna
ipykernel
ipython
ipython-genutils
ipywidgets
itsdangerous
jedi
Jinja2
jmespath
joblib
jsonschema
jupyter
jupyter-client
jupyter-console
jupyter-core
jupyterlab-pygments
jupyterlab-widgets
keras
Keras-Preprocessing
keras-rl2
kiwisolver
lxml
Markdown
MarkupSafe
matplotlib
mistune
murmurhash
mypy-extensions
nbconvert
nbformat
notebook
numpy
oauthlib
opt-einsum
packaging
pandas
pandas-datareader
pandas-gbq
pandocfilters
parso
pbr
pexpect
pickleshare
Pillow
platformdirs
preshed
prometheus-client
prometheus-client
prompt-toolkit
proto-plus
protobuf
ptyprocess
pyarrow
pyasn1
pyasn1-modules
pydata-google-auth
Pygments
pyparsing
pyrsistent
python-dateutil
pytz
PyYAML
pyzmq
qtconsole
QtPy
requests
requests-oauthlib
s3transfer
scikit-learn
scipy
Send2Trash
six
sklearn
smart-open
srsly
stevedore
tensorboard
tensorboard-plugin-wit
tensorflow
tensorflow-estimator
tensorflow-hub
terminado
testpath
thinc
tornado
tqdm
traitlets
typer
typing-extensions
typing-inspect
urllib3
virtualenv-clone
virtualenvwrapper
wasabi
wcwidth
webencodings
Werkzeug
widgetsnbextension
wrapt
"requirements.txt" 142L, 1557C                           140,5         Bot
appnope
astunparse
attrs
backcall
backports.entry-points-selectable
bayesian-optimization
bleach
blis
boto3
botocore
cachetools
catalogue
certifi
cffi
clang
click
cloudpickle
cycler
cymem
debugpy
decorator
defusedxml
distlib
entrypoints
filelock
first
Flask
flatbuffers
gast
gensim
google-api-core
google-auth
google-auth-oauthlib
google-cloud-bigquery
google-cloud-bigquery-storage
google-cloud-core
google-crc32c
google-pasta
google-resumable-media
googleapis-common-protos
grpcio
gym
h5py
idna
ipykernel
ipython
ipython-genutils
ipywidgets
itsdangerous
jedi
Jinja2
                           47,5          Top
prompt-toolkit
proto-plus
protobuf
ptyprocess
pyarrow
pyasn1
pyasn1-modules
pydata-google-auth
Pygments
pyparsing
pyrsistent
python-dateutil
pytz
PyYAML
pyzmq
qtconsole
QtPy
requests
requests-oauthlib
s3transfer
scikit-learn
scipy
Send2Trash
six
sklearn
smart-open
srsly
stevedore
tensorboard
tensorboard-plugin-wit
tensorflow
tensorflow-estimator
tensorflow-hub
terminado
testpath
thinc
tornado
tqdm
traitlets
typer
typing-extensions
typing-inspect
urllib3
virtualenv-clone
virtualenvwrapper
wasabi
wcwidth
webencodings
Werkzeug
widgetsnbextension
wrapt
raz.shmuely@dev-raz-ds-test-1:~/hens$ vim requirements.txt
raz.shmuely@dev-raz-ds-test-1:~/hens$ 0;7;52M0;7;52m0;7;52M0;7;52m^C
raz.shmuely@dev-raz-ds-test-1:~/hens$ pwd
/home/raz.shmuely/hens
raz.shmuely@dev-raz-ds-test-1:~/hens$ 0;23;52M0;1;52m0;48;8M0;48;8m^C
raz.shmuely@dev-raz-ds-test-1:~/hens$ pwd
/home/raz.shmuely/hens
raz.shmuely@dev-raz-ds-test-1:~/hens$ scp /Users/raz.shmuely raz.shmuely@dev-raz-ds-test-1:/home/raz.shmuely/hens/requirements.txt
The authenticity of host 'dev-raz-ds-test-1 (10.6.0.72)' can't be established.
ECDSA key fingerprint is SHA256:Ra4/oAvgIdOkPfNdI8Y3+1VheBMRW4xkptuw+dAPNJg.
Are you sure you want to continue connecting (yes/no)? no
Host key verification failed.
lost connection
raz.shmuely@dev-raz-ds-test-1:~/hens$ scp /Users/raz.shmuely raz.shmuely@dev-raz-ds-test-1:requirements.txt
The authenticity of host 'dev-raz-ds-test-1 (10.6.0.72)' can't be established.
ECDSA key fingerprint is SHA256:Ra4/oAvgIdOkPfNdI8Y3+1VheBMRW4xkptuw+dAPNJg.
Are you sure you want to continue connecting (yes/no)? yes
Warning: Permanently added 'dev-raz-ds-test-1,10.6.0.72' (ECDSA) to the list of known hosts.
raz.shmuely@dev-raz-ds-test-1: Permission denied (publickey).
lost connection
raz.shmuely@dev-raz-ds-test-1:~/hens$ cat requirements.tx
cat: requirements.tx: No such file or directory
raz.shmuely@dev-raz-ds-test-1:~/hens$ cat requirements.txt
absl-py
appnope
astunparse
attrs
backcall
backports.entry-points-selectable
bayesian-optimization
bleach
blis
boto3
botocore
cachetools
catalogue
certifi
cffi
clang
click
cloudpickle
cycler
cymem
debugpy
decorator
defusedxml
distlib
entrypoints
filelock
first
Flask
flatbuffers
gast
gensim
google-api-core
google-auth
google-auth-oauthlib
google-cloud-bigquery
google-cloud-bigquery-storage
google-cloud-core
google-crc32c
google-pasta
google-resumable-media
googleapis-common-protos
grpcio
gym
h5py
idna
ipykernel
ipython
ipython-genutils
ipywidgets
itsdangerous
jedi
Jinja2
jmespath
joblib
jsonschema
jupyter
jupyter-client
jupyter-console
jupyter-core
jupyterlab-pygments
jupyterlab-widgets
keras
Keras-Preprocessing
keras-rl2
kiwisolver
lxml
Markdown
MarkupSafe
matplotlib
mistune
murmurhash
mypy-extensions
nbconvert
nbformat
notebook
numpy
oauthlib
opt-einsum
packaging
pandas
pandas-datareader
pandas-gbq
pandocfilters
parso
pbr
pexpect
pickleshare
Pillow
platformdirs
preshed
prometheus-client
prompt-toolkit
proto-plus
protobuf
ptyprocess
pyarrow
pyasn1
pyasn1-modules
pydata-google-auth
Pygments
pyparsing
pyrsistent
python-dateutil
pytz
PyYAML
pyzmq
qtconsole
QtPy
requests
requests-oauthlib
s3transfer
scikit-learn
scipy
Send2Trash
six
sklearn
smart-open
srsly
stevedore
tensorboard
tensorboard-plugin-wit
tensorflow
tensorflow-estimator
tensorflow-hub
terminado
testpath
thinc
tornado
tqdm
traitlets
typer
typing-extensions
typing-inspect
urllib3
virtualenv-clone
virtualenvwrapper
wasabi
wcwidth
webencodings
Werkzeug
widgetsnbextension
wrapt